requirements.txt
grpcio
grpcio-tools
requests
gradio
Pillow
opencv-python
sqlite-utils
python-dotenv
# whisper-cpp
coqui-tts
pyttsx3  # Fallback TTS engine
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
langdetect>=1.0.9
openai-whisper>=20231117
schedule>=1.2.0  # For session cleanup scheduling
